refactor: hoist per-kind cli logic helpers into lib/baselines/kinds and migrate dependent tests (resolves #1989)
