+++ .agents/scripts/lib/baselines/kinds/maintainability.js
@@ -11,6 +11,16 @@
  * `typhonjs-escomplex` too.
  */
 
+import { readBaselineAtRef } from '../../baseline-loader.js';
+import { loadBaseline } from '../../gates/baseline-store.js';
+import { Logger } from '../../Logger.js';
+import { getBaseline } from '../../maintainability-utils.js';
+import {
+  applyFloorPolicy,
+  formatViolation,
+  loadFloorConfig,
+  parseFloorFlag,
+} from '../../quality-floors.js';
 import { canonicalise } from '../path-canon.js';
 import { mergeRowsByScope } from '../scope.js';
 import { kernelVersion as crapKernelVersion } from './crap.js';
@@ -150,4 +160,128 @@ export function mergeRows(prior, regenerated, scope) {
     scope,
     scopeKey: (row) => row.path,
   });
+}
+
+// ---------------------------------------------------------------------------
+// CLI-facing pure helpers (Story #1981, Task #1989).
+// Hoisted from `.agents/scripts/check-maintainability.js`.
+// ---------------------------------------------------------------------------
+
+// Envelope version for the --json parity output. Bump when the report
+// shape changes so downstream agent workflows can detect breaks without
+// guessing. 1.1.0 — TypeScript support landed in 5.29.0.
+export const MI_REPORT_KERNEL_VERSION = '1.1.0';
+
+/**
+ * Pure: build the MI parity envelope. Shape matches the CRAP `--json`
+ * output: `{ kernelVersion, summary, violations }` sans `fixGuidance`
+ * (MI scores don't decompose along the two CRAP axes).
+ *
+ * @param {Record<string, number>} scores current MI scores keyed by file
+ * @param {{
+ *   regressions?: number,
+ *   newFiles?: number,
+ *   improvements?: number,
+ *   regressedFiles?: Array<{file: string, current: number, baseline: number, drop: number}>
+ * }} stats
+ * @param {{ scope?: 'diff' | 'full', diffRef?: string | null }} [scopeInfo]
+ */
+export function buildMaintainabilityReport(scores, stats, scopeInfo) {
+  const total = Object.keys(scores ?? {}).length;
+  const violations = (stats?.regressedFiles ?? []).map((r) => ({
+    file: r.file,
+    current: r.current,
+    baseline: r.baseline,
+    drop: r.drop,
+    kind: 'regression',
+  }));
+  const scope = scopeInfo?.scope === 'full' ? 'full' : 'diff';
+  const diffRef = scope === 'full' ? null : (scopeInfo?.diffRef ?? null);
+  return {
+    kernelVersion: MI_REPORT_KERNEL_VERSION,
+    summary: {
+      total,
+      regressions: stats?.regressions ?? 0,
+      newFiles: stats?.newFiles ?? 0,
+      improvements: stats?.improvements ?? 0,
+      scope,
+      diffRef,
+    },
+    violations,
+  };
+}
+
+/** Thin wrapper: delegate to baseline-store with MI-specific defaults. */
+export function loadMaintainabilityBaseline({
+  baselinePath,
+  epicRef,
+  readBaseline = getBaseline,
+  readAtRef = readBaselineAtRef,
+  logger = console,
+}) {
+  const parsed = loadBaseline({
+    baselinePath,
+    epicRef,
+    readAtRef,
+    readFromTree: ({ baselinePath: p }) => readBaseline(p),
+    logger,
+    label: 'Maintainability',
+  });
+  // Epic-ref read may return a non-object — coerce to {} so downstream
+  // `Object.entries` / `Object.keys` never throws.
+  if (epicRef && (parsed === null || typeof parsed !== 'object')) return {};
+  // Story #1895: the on-disk baseline switched from the flat
+  // `{ path: mi }` map to the canonical envelope shape. Project envelope
+  // back to the legacy flat shape so downstream comparators keep working.
+  if (
+    parsed &&
+    typeof parsed === 'object' &&
+    !Array.isArray(parsed) &&
+    Array.isArray(parsed.rows) &&
+    typeof parsed.$schema === 'string'
+  ) {
+    const flat = {};
+    for (const row of parsed.rows) {
+      if (row && typeof row.path === 'string' && typeof row.mi === 'number') {
+        flat[row.path] = row.mi;
+      }
+    }
+    return flat;
+  }
+  return parsed;
+}
+
+/**
+ * Story #1602 — absolute MI floor (≥70 by default). Pure decision
+ * helper: returns `0` when the gate is skipped or all files clear the
+ * floor, `1` when one or more files are below floor. The CLI wrapper
+ * translates the return into a `process.exit(1)`; tests can drive the
+ * helper directly and assert the return code.
+ *
+ * `options.floors` overrides the `loadFloorConfig()` default — used by
+ * tests and by callers that have already loaded the floor config.
+ *
+ * Opt-out: pass `--floor=off` in `argv` for baseline-update runs.
+ *
+ * @returns {0 | 1}
+ */
+export function enforceMaintainabilityFloor(scores, argv, options = {}) {
+  if (!parseFloorFlag(argv)) {
+    Logger.info('[Maintainability] ⚠️  floor gate skipped (--floor=off)');
+    return 0;
+  }
+  const floors = options.floors ?? loadFloorConfig();
+  const records = Object.entries(scores).map(([file, mi]) => ({ file, mi }));
+  const { violations } = applyFloorPolicy(records, floors, 'maintainability');
+  if (violations.length === 0) return 0;
+  Logger.error(
+    `[Maintainability] ❌ Absolute MI floor violated (${violations.length} file(s); floor=${floors.maintainability}):`,
+  );
+  for (const v of violations) {
+    Logger.error(`                ${formatViolation(v)}`);
+  }
+  Logger.error(
+    '[Maintainability] Refactor the flagged file(s); the floor is non-negotiable. Use `--floor=off` only when running `maintainability:update`.',
+  );
+  return 1;
 }
